+++ operators/ARMORED_export.py
@@ -1,11 +1,13 @@
-version = (2, 0, 1)
+version = (3, 0, 0)
 
 import bpy
 import os
+import pathlib
 
 
-export_path = 'D:\Desktop'
-export_name = 'temp'       # NOTE: file extension will be added automatically based on the exporter.
+# IF THE ARMORED TOOLKIT IS INSTALLED, YOU CAN NOW CHANGE THE EXPORT PATH IN THE ADDON PREFERENCES.
+
+fallback_path = str(pathlib.Path.home() / 'Desktop')
 
 
 class Export:
@@ -13,21 +15,54 @@ class Export:
 	Abstract class for exporting your viewport selection without any prompts in Blender.
 	'''
  
-	bl_options = {'REGISTER'}
+	bl_options = {'REGISTER', 'UNDO'}
+	bl_property = 'export_path'
 	extension = NotImplemented
 
+	file_name : bpy.props.StringProperty(
+		name='File Name', default='')
+
+	dir_path_override : bpy.props.StringProperty(
+		name='File Path', default='', subtype='DIR_PATH', options={'SKIP_SAVE'})
+
+	def invoke(self, context, event):		
+		return context.window_manager.invoke_props_dialog(self)
+
 	def execute(self, context):
+		if self.dir_path_override:
+			print('Using override path')
+			self.export_path = self.dir_path_override
+
+		else:
+			
+			try:
+				self.export_path = self._get_path_from_addon(context)
+				print('Using addon path')
+
+			except Exception as e:
+				print(e)
+				print(f'Could not get path from ARMORED-Toolkit, using fallback path \'{fallback_path}\'')
+				self.export_path = fallback_path
+			
+			print(f'Export Path is {self.export_path}')
+
 		self._validate_extension()
 
 		if self._nothing_selected(context):
 			self._report_nothing_selected()
 			return {'CANCELLED'}
 
-		self.export_path = self._validate_path(export_path)
-		self.file_path = os.path.join(self.export_path, export_name + self.extension)
+		if self._path_does_not_exist(self.export_path):
+			self._report_path_does_not_exist()
+			return {'CANCELLED'}
+
+		if not self.file_name:
+			self.file_name = 'temp'
+
+		self.file_path = os.path.join(self.export_path, self.file_name + self.extension)
 
 		self._run_exporter()
-		self._report_finished()
+		self._report_finished(context)
 
 		return {'FINISHED'}
 	
@@ -35,15 +70,21 @@ class Export:
 	def _run_exporter(self):
 		raise NotImplementedError
 	
+	def _get_path_from_addon(self, context):
+		armored_toolkit = context.preferences.addons['ARMORED-Toolkit']
+		props = armored_toolkit.preferences
+
+		return getattr(props, 'export_path')
+	
 
 	# VALIDATION >>
 	
 	def _nothing_selected(self, context):
 		return not context.selected_objects
 
-	def _validate_path(self, path):
-		return path if os.path.exists(path) else os.path.expanduser("~/Desktop")
-	
+	def _path_does_not_exist(self, path):
+		return not os.path.exists(path)
+			
 	def _validate_extension(self):
 		if self.extension is NotImplemented:
 			raise NotImplementedError(
@@ -54,39 +95,46 @@ class Export:
 
 	def _report_nothing_selected(self):
 		self.report({'WARNING'}, 'Select the objects you want to export')
+	
+	def _report_path_does_not_exist(self):
+		self.report({'ERROR'}, f'Path \'{self.export_path}\' does not exist. Check the <Fast Export> path in the Blender file path preferences.')
 
-	def _report_finished(self):
-		self.report({'INFO'}, f'Exported {export_name + self.extension} to {self.export_path}')
+	def _report_finished(self, context):
+		self.report({'INFO'}, f"Exported '{self.file_name}' to '{self.export_path}'")
 
 
 
 class ARMORED_OT_export_OBJ(bpy.types.Operator, Export):
 	'''Export selected as OBJ.
 
-www.armoredColony.com '''
+	armoredColony.com '''
  
 	bl_idname = 'object.armored_export_obj'
 	bl_label = 'ARMORED Export OBJ'
 	extension = '.obj'
 
+	material_split : bpy.props.BoolProperty(name='Material Split', default=False)
+
 	def _run_exporter(self):
 		bpy.ops.wm.obj_export(
 			filepath=self.file_path,
 			export_selected_objects=True,
+			global_scale=1,
 			export_uv=False,
 			export_normals=False,
 			apply_modifiers=True,
 			export_materials=True,
 			export_pbr_extensions=True,
+			# export_object_groups=not self.material_split,
 			export_object_groups=False,
-			export_material_groups=True,
+			export_material_groups=self.material_split,
 		)
 
 
 class ARMORED_OT_export_FBX(bpy.types.Operator, Export):
 	'''Export selected as FBX.
 
-www.armoredColony.com '''
+	armoredColony.com '''
  
 	bl_idname = 'object.armored_export_fbx'
 	bl_label = 'ARMORED Export FBX'
@@ -96,9 +144,17 @@ www.armoredColony.com '''
 		bpy.ops.export_scene.fbx(
 			filepath=self.file_path, 
 			use_selection=True, 
-			bake_space_transform=True,
-			mesh_smooth_type='FACE',
-			bake_anim=True,
+			# bake_space_transform=True,
+			# mesh_smooth_type='FACE',
+			bake_anim=False,
+			object_types={
+				# 'ARMATURE', 
+				# 'CAMERA', 
+				# 'EMPTY', 
+				# 'LIGHT', 
+				'MESH', 
+				# 'OTHER'
+				}
 		)