+++ utils/icons.py
@@ -1,4 +1,4 @@
-import bpy.utils.previews
+import bpy
 import os
 
 from . import paths